renamed generator class files
