+++ lib/commands/generate/index.js
@@ -1,7 +1,7 @@
 const chalk = require("chalk");
 const path = require("path");
-const StyleGenerator = require("./style");
-const ScriptGenerator = require("./script");
+const StyleGenerator = require("./style/generator");
+const ScriptGenerator = require("./script/generator");
 
 module.exports = async (args, options, logger) => {
   const { type, name } = args;
