+++ lib/commands/generate/script/generator.js
