+++ lib/commands/generate/style/generator.js
